// src/types/teachers/add.tsx
import { TeacherData } from './list';
import { TeachersListStatus } from '../../enums/teachers/list';

export interface TeachersAddPayload {
    personel_id: number;
    name_surname: string;
    short_name: string;
    branch: string;
    class_teacher_id: number;
    social_club: string;
    email: string;
}

export interface TeachersAddState {
    data: TeacherData | null;
    status: TeachersListStatus;
    error: string | null;
}